+++ RELEASES.md
@@ -1,4 +1,14 @@
 ## 0.0.8-SNAPSHOT/20210402-19:45:31
+
+- Issues fixed:
+  - ebRigol (no device found): _version command does not work
+
+- Features added
+  - HELLO.org w. hello.py && hello2.py
+  - hidden menu item: menu command starting with _-char
+
+
+
 ## 0.0.7/20210402-19:44:32
 
 - From from https://github.com/jarjuk/UTG900 0.0.6-SNAPSHOT